'抽出lib'
