+++ views/message.py
@@ -5,9 +5,11 @@ import json
 
 from flask import request, g, jsonify, session
 
-from models.message import Message
 from models.user import User
-from views.base import validate_user_login, allow_cross_domain, get_timestamp
+from models.message import Message
+
+from lib.utils import get_timestamp
+from lib.decorators import validate_user_login, allow_cross_domain
 
 
 @allow_cross_domain
